hypre solver solve a-term.
Change atoi to atof when parsing epsp and epsm.
Helper.cpp: print_interface_info, print min distance from gridpoint to interface.

+++ global.cpp
@@ -173,11 +173,11 @@ static struct option long_options[] = {
             break;
 
          case 'M':
-            EPSILONM = atoi(optarg);
+            EPSILONM = atof(optarg);
          break;
 
          case 'P':
-            EPSILONP = atoi(optarg);
+            EPSILONP = atof(optarg);
          break;
 
          case 'r':
